renamed component js files to index.js. Added nav to app.js, now displays on all pages. Deleted home buttons for login and signup.
