+++ client/src/components/Nav/index.js
@@ -4,15 +4,14 @@ import './style.css'
 
 function Nav(props) {
     return (
-        <div className='float-right'>
+        <div id='backgroundColor'>
             <nav className="navbar navbar-expand-lg navbar-light bg-light">
                 <button className="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
                     <span className="navbar-toggler-icon"></span>
                 </button>
                 <div className="collapse navbar-collapse" id="navbarNav">
                     <ul className="navbar-nav">
-                    <li className={( props.loggedIn === "true") ? "" : "hiddenForNav"}><Link to="#">{`Welcome, ${localStorage.getItem('name')}!`}</Link></li>
-                    <li onClick={props.patientLogin}>
+                        <li onClick={props.patientLogin}>
                             <Link
                                 to='/'
                                 className={
@@ -84,6 +83,7 @@ function Nav(props) {
                                 Sign Up
                             </Link>
                         </li>
+                        <li id='loginName' className={(props.loggedIn === "true") ? "" : "hiddenForNav"}><Link to="#">{`Welcome, ${localStorage.getItem('name')}!`}</Link></li>
                     </ul>
                 </div>
             </nav>
