+++ client/src/pages/SignUp.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
-import PatientSignup from "../components/Signup/signup";
-import Nav from '../components/Nav';
+import PatientSignup from "../components/Signup";
 import API from '../utils/API';
